requirements.txt
pandas>=2.1,<2.3
pyarrow==16.1.0
httpx==0.27.0
tenacity==8.5.0
typer==0.12.3
rich==13.7.1
python-dotenv==1.0.1
pytest==8.2.0
matplotlib==3.8.4
